backendproject2/src/main/java/com/models/User.java
package com.models;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

@Entity
@Table(name="User_127")
public class User {
	 	@Id
	 	private String email;
	    private String password;
	    private String firstName;
	    private String lastName;
	    private long contactNumber;
	    private String role;
	    private String onlineStatus;
		public long getContactNumber() {
			return contactNumber;
		}
		public void setContactNumber(long contactNumber) {
			this.contactNumber = contactNumber;
		}
		public String getOnlineStatus() {
			return onlineStatus;
		}
		public void setOnlineStatus(String onlineStatus) {
			this.onlineStatus = onlineStatus;
		}
		public String getRole() {
			return role;
		}
		public void setRole(String role) {
			this.role = role;
		}
		public String getLastName() {
			return lastName;
		}
		public void setLastName(String lastName) {
			this.lastName = lastName;
		}
		public String getFirstName() {
			return firstName;
		}
		public void setFirstName(String firstName) {
			this.firstName = firstName;
		}
		public String getEmail() {
			return email;
		}
		public void setEmail(String email) {
			this.email = email;
		}
		public String getPassword() {
			return password;
		}
		public void setPassword(String password) {
			this.password = password;
		}
	     
}
